requirements.txt
langchain==0.1.12
langchain-community==0.0.29
faiss-cpu==1.7.4
streamlit==1.32.2
pypdf==5.3.0
numpy==1.26.4
pillow==10.2.0 